upfate

+++ src/components/tabs/autobuyers/DimensionBoostAutobuyerBox.js
@@ -97,7 +97,7 @@ export default {
       </label>
       <AutobuyerInput
         :autobuyer="autobuyer"
-        type="int"
+        type="decimal"
         property="maxDimBoosts"
       />
     </template>
@@ -140,7 +140,7 @@ export default {
       </label>
       <AutobuyerInput
         :autobuyer="autobuyer"
-        type="int"
+        type="decimal"
         property="galaxies"
       />
     </template>
