+++ src/components/tabs/autobuyers/GalaxyAutobuyerBox.js
@@ -83,7 +83,7 @@ export default {
       </label>
       <AutobuyerInput
         :autobuyer="autobuyer"
-        type="int"
+        type="decimal"
         property="maxGalaxies"
       />
     </template>
